+++ meta-memfault/recipes-memfault/memfaultd/files/memfaultd/src/plugins/reboot/reboot.c
@@ -23,6 +23,7 @@
 #include "memfault/core/math.h"
 #include "memfault/core/reboot_reason_types.h"
 #include "memfault/util/linux_boot_id.h"
+#include "memfault/util/reboot_reason.h"
 #include "memfaultd.h"
 #include "reboot_last_boot_id.h"
 #include "reboot_process_pstore.h"
@@ -38,65 +39,6 @@ typedef struct MemfaultdRebootReasonSource {
   bool (*read_and_clear)(sMemfaultd *memfaultd, eMemfaultRebootReason *reboot_reason);
   const char *name;
 } sMemfaultdRebootReasonSource;
-
-static const char *prv_reboot_reason_str(eMemfaultRebootReason reboot_reason) {
-  switch (reboot_reason) {
-    case kMfltRebootReason_Unknown:
-      return "Unknown";
-    case kMfltRebootReason_UserShutdown:
-      return "UserShutdown";
-    case kMfltRebootReason_UserReset:
-      return "UserReset";
-    case kMfltRebootReason_FirmwareUpdate:
-      return "FirmwareUpdate";
-    case kMfltRebootReason_LowPower:
-      return "LowPower";
-    case kMfltRebootReason_DebuggerHalted:
-      return "DebuggerHalted";
-    case kMfltRebootReason_ButtonReset:
-      return "ButtonReset";
-    case kMfltRebootReason_PowerOnReset:
-      return "PowerOnReset";
-    case kMfltRebootReason_SoftwareReset:
-      return "SoftwareReset";
-    case kMfltRebootReason_DeepSleep:
-      return "DeepSleep";
-    case kMfltRebootReason_PinReset:
-      return "PinReset";
-    case kMfltRebootReason_UnknownError:
-      return "UnknownError";
-    case kMfltRebootReason_Assert:
-      return "Assert";
-    case kMfltRebootReason_WatchdogDeprecated:
-      return "WatchdogDeprecated";
-    case kMfltRebootReason_BrownOutReset:
-      return "BrownOutReset";
-    case kMfltRebootReason_Nmi:
-      return "Nmi";
-    case kMfltRebootReason_HardwareWatchdog:
-      return "HardwareWatchdog";
-    case kMfltRebootReason_SoftwareWatchdog:
-      return "SoftwareWatchdog";
-    case kMfltRebootReason_ClockFailure:
-      return "ClockFailure";
-    case kMfltRebootReason_KernelPanic:
-      return "KernelPanic";
-    case kMfltRebootReason_FirmwareUpdateError:
-      return "FirmwareUpdateError";
-    case kMfltRebootReason_BusFault:
-      return "BusFault";
-    case kMfltRebootReason_MemFault:
-      return "MemFault";
-    case kMfltRebootReason_UsageFault:
-      return "UsageFault";
-    case kMfltRebootReason_HardFault:
-      return "HardFault";
-    case kMfltRebootReason_Lockup:
-      return "Lockup";
-    default:
-      return "???";
-  }
-}
 
 /**
  * @brief Builds event JSON object for posting to events API
@@ -388,11 +330,12 @@ static eMemfaultRebootReason prv_resolve_reboot_reason(sMemfaultd *memfaultd, co
         prioritized_reason = reason;
         prioritized_reason_set = true;
         fprintf(stderr, "reboot:: Using reboot reason %s (0x%04x) from %s source for boot_id %s\n",
-                prv_reboot_reason_str(reason), reason, s_reboot_reason_sources[i].name, boot_id);
+                memfaultd_reboot_reason_str(reason), reason, s_reboot_reason_sources[i].name,
+                boot_id);
       } else {
-        fprintf(stderr,
-                "reboot:: Discarded reboot reason %s (0x%04x) from %s source for boot_id %s\n",
-                prv_reboot_reason_str(reason), reason, s_reboot_reason_sources[i].name, boot_id);
+        fprintf(
+          stderr, "reboot:: Discarded reboot reason %s (0x%04x) from %s source for boot_id %s\n",
+          memfaultd_reboot_reason_str(reason), reason, s_reboot_reason_sources[i].name, boot_id);
       }
     }
   }