+++ test-scripts/02_memfaultd_enable_data_collection_test.py
@@ -13,10 +13,10 @@ def test_start(qemu: QEMU):
     qemu.systemd_wait_for_service_state("memfaultd.service", "active")
 
     qemu.exec_cmd("memfaultd --enable-data-collection")
-    qemu.child().expect("Data collection state already set")
+    qemu.child().expect("Data collection is already enabled.")
 
     qemu.exec_cmd("memfaultd --disable-data-collection")
-    qemu.child().expect("Disabling data collection")
+    qemu.child().expect("Disabling data collection.")
 
     # Work-around for checking the state in the next line before memfaultd has even restarted:
     # FIXME: track the journal logs instead of using `systemctl is-active`
@@ -25,7 +25,44 @@ def test_start(qemu: QEMU):
     qemu.systemd_wait_for_service_state("memfaultd.service", "active")
 
     qemu.exec_cmd("memfaultd --disable-data-collection")
-    qemu.child().expect("Data collection state already set")
+    qemu.child().expect("Data collection is already disabled.")
+
+    # Check that the service restarted:
+    qemu.exec_cmd("journalctl -u memfaultd.service")
+    qemu.child().expect("Stopped memfaultd daemon")
+    qemu.child().expect("Starting memfaultd daemon")
+
+    # Check that the service did not fail:
+    # FIXME: pexpect does not have a "not expecting" API :/
+    qemu.exec_cmd("journalctl -u memfaultd.service")
+    output = b""
+    while True:
+        try:
+            output += qemu.child().read_nonblocking(size=1024, timeout=0.5)
+        except pexpect.TIMEOUT:
+            break
+    assert b"memfaultd.service: Scheduled restart job" not in output
+
+
+def test_via_memfaultctl(qemu: QEMU):
+    qemu.exec_cmd("memfaultctl enable-data-collection")
+    qemu.child().expect("Enabling data collection.")
+    qemu.systemd_wait_for_service_state("memfaultd.service", "active")
+
+    qemu.exec_cmd("memfaultctl enable-data-collection")
+    qemu.child().expect("Data collection is already enabled.")
+
+    qemu.exec_cmd("memfaultctl disable-data-collection")
+    qemu.child().expect("Disabling data collection.")
+
+    # Work-around for checking the state in the next line before memfaultd has even restarted:
+    # FIXME: track the journal logs instead of using `systemctl is-active`
+    time.sleep(0.5)
+
+    qemu.systemd_wait_for_service_state("memfaultd.service", "active")
+
+    qemu.exec_cmd("memfaultctl disable-data-collection")
+    qemu.child().expect("Data collection is already disabled.")
 
     # Check that the service restarted:
     qemu.exec_cmd("journalctl -u memfaultd.service")