+++ test-scripts/11_write_attributes_test.py
@@ -1,0 +1,53 @@
+#
+# Copyright (c) Memfault, Inc.
+# See License.txt for details
+from memfault_service_tester import MemfaultServiceTester
+from qemu import QEMU
+
+
+# Assumptions:
+# - The machine/qemu is built with a valid project key of a project on app.memfault.com,
+#   or whatever the underlying QEMU instance points at.
+# - The MEMFAULT_E2E_* environment variables are set to match whatever the underlying
+#   QEMU instance points at.
+def test(
+    qemu: QEMU, memfault_service_tester: MemfaultServiceTester, qemu_device_id: str
+):
+    qemu.exec_cmd("memfaultctl enable-data-collection")
+
+    # Start following memfaultd logs
+    qemu.exec_cmd("journalctl -u memfaultd -n 0 -f &")
+
+    # Wait until memfaultd has started
+    qemu.child().expect("Started memfaultd daemon.")
+
+    # Write attributes
+    qemu.exec_cmd(
+        'memfaultctl write-attributes a_string=running a_bool=false a_boolish_string=\\"true\\" a_float=42.42'
+    )
+
+    # Poke memfaultd to upload now
+    qemu.exec_cmd("memfaultctl sync")
+
+    # Wait until we have received attributes
+    def _check():
+        attributes = memfault_service_tester.list_attributes(
+            device_serial=qemu_device_id
+        )
+
+        assert attributes
+
+        d = {
+            a["custom_metric"]["string_key"]: a["state"]["value"]
+            for a in attributes
+            if a["state"] is not None
+        }
+
+        assert d["a_string"] == "running"
+        assert d["a_bool"] is False
+        assert d["a_boolish_string"] == "true"
+        assert d["a_float"] == 42.42
+
+    memfault_service_tester.poll_until_not_raising(
+        _check, timeout_seconds=60, poll_interval_seconds=1
+    )